requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers
faiss-cpu
fastapi
uvicorn
rank_bm25
rapidfuzz
boto3
python-dotenv